# Speed up fetching/loading of extensions via `graphql?Extensions`