--- a/client/shared/src/extensions/helpers.test.ts
+++ b/client/shared/src/extensions/helpers.test.ts
@@ -5,12 +5,16 @@ import { TestScheduler } from 'rxjs/testing'
 import { ErrorGraphQLResult, SuccessGraphQLResult } from '../graphql/graphql'
 import { PlatformContext } from '../platform/context'
 
-import { ConfiguredExtension } from './extension'
+import { ConfiguredExtension, ConfiguredExtensionManifestDefaultFields } from './extension'
 import { ExtensionManifest } from './extensionManifest'
 import { queryConfiguredRegistryExtensions } from './helpers'
 
 const TEST_MANIFEST_RAW = '{"publisher":"a","url":"https://example.com","activationEvents":[]}'
-const TEST_MANIFEST: ExtensionManifest = { publisher: 'a', url: 'https://example.com', activationEvents: [] }
+const TEST_MANIFEST: Pick<ExtensionManifest, ConfiguredExtensionManifestDefaultFields | 'publisher'> = {
+    publisher: 'a',
+    url: 'https://example.com',
+    activationEvents: [],
+}
 
 const scheduler = (): TestScheduler => new TestScheduler((actual, expected) => expect(actual).toStrictEqual(expected))
 
@@ -22,7 +26,7 @@ describe('queryConfiguredRegistryExtensions', () => {
                 data: {
                     extensionRegistry: {
                         extensions: {
-                            nodes: [{ extensionID: 'a/b', manifest: { raw: TEST_MANIFEST_RAW } }],
+                            nodes: [{ extensionID: 'a/b', manifest: { jsonFields: TEST_MANIFEST } }],
                         },
                     },
                 },
@@ -34,10 +38,10 @@ describe('queryConfiguredRegistryExtensions', () => {
         })
     })
 
-    it('gets extensions from GraphQL servers not supporting extensions(extensionIDs) and only supporting prioritizeExtensionIDs', () => {
+    it('gets extensions from GraphQL servers not supporting extensions(extensionIDs)/jsonFields and only supporting prioritizeExtensionIDs', () => {
         let calledWithExtensionIDsParameter = false
         const requestGraphQL: PlatformContext['requestGraphQL'] = ({ request }) => {
-            if (request.includes('prioritizeExtensionIDs: ')) {
+            if (request.includes('prioritizeExtensionIDs: ') && !request.includes('jsonFields(')) {
                 // eslint-disable-next-line @typescript-eslint/consistent-type-assertions
                 return of({
                     data: {
@@ -56,6 +60,9 @@ describe('queryConfiguredRegistryExtensions', () => {
                 errors: [
                     {
                         message: 'Unknown argument "extensionIDs" on field "extensions" of type "ExtensionRegistry".',
+                    },
+                    {
+                        message: 'Cannot query field "jsonFields" on type "ExtensionManifest".',
                     },
                 ] as GraphQLError[],
             } as ErrorGraphQLResult)