--- a/cmd/frontend/graphqlbackend/extension_registry.go
+++ b/cmd/frontend/graphqlbackend/extension_registry.go
@@ -128,6 +128,7 @@ type RegistryExtension interface {
 // ExtensionManifest is the interface for the GraphQL type ExtensionManifest.
 type ExtensionManifest interface {
 	Raw() string
+	JSONFields(*struct{ Fields []string }) JSONValue
 	Description() (*string, error)
 	BundleURL() (*string, error)
 }