--- a/cmd/frontend/registry/api/extension_manifest.go
+++ b/cmd/frontend/registry/api/extension_manifest.go
@@ -1,6 +1,7 @@
 package api
 
 import (
+	"encoding/json"
 	"sync"
 
 	"github.com/sourcegraph/sourcegraph/cmd/frontend/graphqlbackend"
@@ -35,6 +36,30 @@ func (r *extensionManifest) parse() (*schema.SourcegraphExtensionManifest, error
 }
 
 func (r *extensionManifest) Raw() string { return r.raw }
+
+func (r *extensionManifest) JSONFields(args *struct{ Fields []string }) graphqlbackend.JSONValue {
+	return jsonValueWithFields(r.raw, args.Fields)
+}
+
+func jsonValueWithFields(jsoncStr string, fields []string) graphqlbackend.JSONValue {
+	o := map[string]json.RawMessage{}
+	jsonc.Unmarshal(jsoncStr, &o) // ignore error and treat as empty object
+
+	keepField := func(field string) bool {
+		for _, f := range fields {
+			if f == field {
+				return true
+			}
+		}
+		return false
+	}
+	for field := range o {
+		if !keepField(field) {
+			delete(o, field)
+		}
+	}
+	return graphqlbackend.JSONValue{Value: o}
+}
 
 func (r *extensionManifest) Description() (*string, error) {
 	parsed, err := r.parse()