--- a/enterprise/cmd/frontend/internal/registry/extensions_db.go
+++ b/enterprise/cmd/frontend/internal/registry/extensions_db.go
@@ -232,6 +232,7 @@ type dbExtensionsListOptions struct {
 	Query                  string // matches the extension ID and latest release's manifest's title
 	Category               string // matches the latest release's manifest's categories array
 	Tag                    string // matches the latest release's manifest's tags array
+	ExtensionIDs           []string
 	PrioritizeExtensionIDs []string
 	*database.LimitOffset
 }
@@ -273,6 +274,14 @@ func (o dbExtensionsListOptions) sqlConditions() []*sqlf.Query {
 	}
 	if o.Tag != "" {
 		conds = append(conds, sqlf.Sprintf(`CASE WHEN rer.manifest IS NOT NULL THEN (rer.manifest->>'tags')::jsonb @> to_json(%s::text)::jsonb ELSE false END`, o.Tag))
+	}
+	if o.ExtensionIDs != nil {
+		ids := make([]*sqlf.Query, len(o.ExtensionIDs)+1)
+		for i, id := range o.ExtensionIDs {
+			ids[i] = sqlf.Sprintf("%v", id)
+		}
+		ids[len(o.ExtensionIDs)] = sqlf.Sprintf("NULL")
+		conds = append(conds, sqlf.Sprintf(extensionIDExpr+` IN (%v)`, sqlf.Join(ids, ",")))
 	}
 	if len(conds) == 0 {
 		conds = append(conds, sqlf.Sprintf("TRUE"))