… client/web/src/integration/blob-viewer.test.ts
… client/web/src/integration/repository.test.ts
… client/web/src/settings/configuration.test.ts
… cmd/frontend/registry/api/extension_connection_graphql_test.go
… cmd/frontend/registry/api/extension_manifest_test.go
… enterprise/cmd/frontend/internal/registry/extension_connection_graphql_test.go
… enterprise/cmd/frontend/internal/registry/extensions_db_test.go